+++ src/components/Customnodes/CustomCloudNode.js
@@ -1,9 +1,6 @@
-// import { Handle, Position } from "reactflow";
 import azure from "../../assets/Azure.png";
 import aws from "../../assets/aws.png";
 import minikube from "../../assets/mini.png";
-
-// const handleStyle = { left: 25 };
 
 function CustomCloudNode({ data }) {
   const cloudProvider = data.data.cloudProvider;
