+++ src/components/Customnodes/CustomImageNode.js
@@ -3,7 +3,6 @@ import postgres from "../../assets/postgresql.png";
 import sql from "../../assets/mongo.png";
 
 function CustomImageNode({ data, isConnectable, selected }) {
-
   return (
     <>
       <NodeResizer
@@ -19,10 +18,30 @@ function CustomImageNode({ data, isConnectable, selected }) {
           alt="database"
         />
       </div>
-      <Handle position={Position.Left} id="target.Left" type="target" isConnectable={isConnectable}/>
-      <Handle position={Position.Top} id="target.Top" type="target" isConnectable={isConnectable}/>
-      <Handle position={Position.Bottom} id="target.Bottom" type="target" isConnectable={isConnectable}/>
-      <Handle position={Position.Right} id="target.Right" type="target" isConnectable={isConnectable}/>
+      <Handle
+        position={Position.Left}
+        id="target.Left"
+        type="target"
+        isConnectable={isConnectable}
+      />
+      <Handle
+        position={Position.Top}
+        id="target.Top"
+        type="target"
+        isConnectable={isConnectable}
+      />
+      <Handle
+        position={Position.Bottom}
+        id="target.Bottom"
+        type="target"
+        isConnectable={isConnectable}
+      />
+      <Handle
+        position={Position.Right}
+        id="target.Right"
+        type="target"
+        isConnectable={isConnectable}
+      />
     </>
   );
 }
