+++ src/components/Customnodes/CustomMessageBrokerNode.js
@@ -1,20 +1,12 @@
-// import { Handle, Position } from "reactflow";
 import kafka from "../../assets/kafka.png";
 import pulsar from "../../assets/pulsar.png";
 import rabbitmq from "../../assets/rabbitmq.png";
-
-// const handleStyle = { left: 25 };
 
 function CustomMessageBrokerNode({ data, isConnectable }) {
   const messageBroker = data.messageBroker;
 
   return (
     <div>
-      {/* <Handle
-        type="target"
-        position={Position.Top}
-        isConnectable={isConnectable}
-      /> */}
       <div>
         {messageBroker === "Kafka" ? (
           <img width="50px" name={messageBroker} src={kafka} alt="kafka" />
@@ -29,12 +21,6 @@ function CustomMessageBrokerNode({ data, isConnectable }) {
           />
         )}
       </div>
-      {/* <Handle
-        type="source"
-        position={Position.Bottom}
-        style={handleStyle}
-        isConnectable={isConnectable}
-      /> */}
     </div>
   );
 }
