+++ src/components/Customnodes/GroupNode.js
@@ -51,4 +51,3 @@ export default function GroupNode({ id, data, selected }) {
     </>
   );
 }
-
