+++ src/components/Customnodes/ResizeableNode.js
@@ -51,4 +51,3 @@ export default function ResizableNode({ id, data, selected }) {
     </>
   );
 }
-
