+++ src/components/Modal/DeleteModal.js
@@ -9,12 +9,12 @@ import {
   AlertDialogOverlay,
 } from "@chakra-ui/react";
 
-const DeleteModal = ({onSubmit,onClose, id, name}) => {
-  const isOpen=true;
-  const cancelRef = React.useRef()
+const DeleteModal = ({ onSubmit, onClose, id, name }) => {
+  const isOpen = true;
+  const cancelRef = React.useRef();
   return (
-  <>
-     <AlertDialog
+    <>
+      <AlertDialog
         isOpen={isOpen}
         leastDestructiveRef={cancelRef}
         onClose={onClose}
@@ -22,27 +22,28 @@ const DeleteModal = ({onSubmit,onClose, id, name}) => {
       >
         <AlertDialogOverlay>
           <AlertDialogContent>
-            <AlertDialogHeader fontSize='lg' fontWeight='bold'>
-              Delete 
+            <AlertDialogHeader fontSize="lg" fontWeight="bold">
+              Delete
             </AlertDialogHeader>
 
             <AlertDialogBody>
-              Are you sure? You want to delete project <span style={{fontWeight:"bold"}}>{name}</span>
+              Are you sure? You want to delete project{" "}
+              <span style={{ fontWeight: "bold" }}>{name}</span>
             </AlertDialogBody>
 
             <AlertDialogFooter>
               <Button ref={cancelRef} onClick={onClose}>
                 Cancel
               </Button>
-              <Button colorScheme='red' onClick={()=>onSubmit({id})} ml={3}>
+              <Button colorScheme="red" onClick={() => onSubmit({ id })} ml={3}>
                 Delete
               </Button>
             </AlertDialogFooter>
           </AlertDialogContent>
         </AlertDialogOverlay>
       </AlertDialog>
-  </>
+    </>
   );
-}
+};
 
 export default DeleteModal;
