+++ src/components/Modal/DeploymentModal.js
@@ -49,7 +49,6 @@ const DeploymentModal = ({
           <ModalCloseButton onClick={handleContainerClose} />
           <ModalBody
             style={{
-              maxHeight: "calc(100vh - 200px)",
               overflowY: "auto",
               maxHeight: "600px",
             }}
