+++ src/components/Modal/UIModal.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useId } from "react";
+import React, { useState, useEffect } from "react";
 import {
   Modal,
   ModalContent,
