+++ src/components/Pages/Designer/Designer.js
@@ -10,29 +10,29 @@ import ReactFlow, {
 import "reactflow/dist/style.css";
 import { Button } from "@chakra-ui/react";
 import { ArrowRightIcon } from "@chakra-ui/icons";
-import Sidebar from "./../components/Sidebar";
+import Sidebar from "../Sidebar/Sidebar";
 import { saveAs } from "file-saver";
-import ServiceModal from "../components/Modal/ServiceModal";
-import UiDataModal from "../components/Modal/UIModal";
-import GroupDataModal from "../components/Modal/GroupDataModel";
-import CustomImageNode from "./Customnodes/CustomImageNode";
-import CustomServiceNode from "./Customnodes/CustomServiceNode";
-import CustomIngressNode from "./Customnodes/CustomIngressNode";
-import CustomAuthNode from "./Customnodes/CustomAuthNode";
-import CustomMessageBrokerNode from "./Customnodes/CustomMessageBrokerNode";
-import CustomCloudNode from "./Customnodes/CustomCloudNode";
-import CustomLoadNode from "./Customnodes/CustomLoadNode";
-import CustomLocalenvironmentNode from "./Customnodes/CustomLocalenvironmentNode";
-import AlertModal from "../components/Modal/AlertModal";
-import resizeableNode from "./Customnodes/ResizeableNode";
-import groupNode from "./Customnodes/GroupNode";
+import ServiceModal from "../../Modal/ServiceModal";
+import UiDataModal from "../../Modal/UIModal";
+import GroupDataModal from "../../Modal/GroupDataModel";
+import CustomImageNode from "../../Customnodes/CustomImageNode";
+import CustomServiceNode from "../../Customnodes/CustomServiceNode";
+import CustomIngressNode from "../../Customnodes/CustomIngressNode";
+import CustomAuthNode from "../../Customnodes/CustomAuthNode";
+import CustomMessageBrokerNode from "../../Customnodes/CustomMessageBrokerNode";
+import CustomCloudNode from "../../Customnodes/CustomCloudNode";
+import CustomLoadNode from "../../Customnodes/CustomLoadNode";
+import CustomLocalenvironmentNode from "../../Customnodes/CustomLocalenvironmentNode";
+import AlertModal from "../../Modal/AlertModal";
+import resizeableNode from "../../Customnodes/ResizeableNode";
+import groupNode from "../../Customnodes/GroupNode";
 import { useLocation } from "react-router-dom";
 import { useHistory } from "react-router-dom";
-import "./../App.css";
-import EdgeModal from "../components/Modal/EdgeModal";
+import "../../../App.css";
+import EdgeModal from "../../Modal/EdgeModal";
 import { useKeycloak } from "@react-keycloak/web";
 import { FiUploadCloud } from "react-icons/fi";
-import ActionModal from "../components/Modal/ActionModal";
+import ActionModal from "../../Modal/ActionModal";
 
 let service_id = 1;
 let database_id = 1;
