+++ src/components/Pages/Sidebar/Sidebar.js
@@ -1,14 +1,12 @@
 import React, { useEffect, useState } from "react";
-import db1 from "../assets/postgresql.png";
-import db2 from "../assets/mongo.png";
-import eurkea from "../assets/eureka.png";
-import keycloakIcon from "../assets/keycloak.png";
-import eck from "../assets/eck.png";
-// import mini from "../assets/mini.png";
-// import docker from "../assets/docker.png";
-import "./../App.css";
+import db1 from "../../../assets/postgresql.png";
+import db2 from "../../../assets/mongo.png";
+import eurkea from "../../../assets/eureka.png";
+import keycloakIcon from "../../../assets/keycloak.png";
+import eck from "../../../assets/eck.png";
+import "../.././../App.css";
 import { Input, FormLabel, Button, Checkbox } from "@chakra-ui/react";
-import DeployModal from "./Modal/DeployModal";
+import DeployModal from "../../Modal/DeployModal";
 import { useKeycloak } from "@react-keycloak/web";
 import { ArrowRightIcon, ArrowLeftIcon } from "@chakra-ui/icons";
 import { useLocation } from "react-router-dom";
